Reorganize some headers
